Make Vm::Nexus Handle the Destroy when Host is not assigned
In the wait_vm_removal_from_load_balancer, we make a call to the host
for dropping the network, which fails when the VM is destroyed before
it gets a host, such as in out-of-capacity situations. This happens in
Kubernetes workflow where the VM is added to a LB before it's creation
is completed.