Fix PVC recreation issue
Before this commit, we were recreating the PVC multiple times and
it was not idempotent. On certain cases it break the flow and pvc
would get stuck in a bad state.

With this change, by using pvc uid, we make sure we delete the pvc
once and if we are not the one recreating the pvc, we make sure we
have added the annotation to let the controller on the destination
node do the migration

Some changes were made to the kubernetes client to patch resources
instead of updating them which is simpler and is idempotent