Add Prog::Base.current_prog to find prog information in the stack
It'd be helpful for my Sshable session conflict assessment to know
what progs and labels are involved in contention, but as the relevant
code is down-stack, it's not immediately obvious how to get access to
the current prog and label information.

Searching the stack is a way to do the job.