Work around bug in endpoint for single server minio clusters
With a single server minio cluster, the endpoint returned by
`JSON.parse(minio_client.admin_info.body)["servers"]` is the cluster
name (no 0 suffix) and not the server name (with 0 suffix). The
purpose of the code is to find the related server, and if there
is only a single server, it is obviously the same one.

This is only a work around. An actual bug fix would be to have
minio return the correct server name. However, I'm not sure how
to do that.