Validate minio cluster information before use
This prevents setting up clusters where minio would fail to start
on a server because one of the per-server values was 0. For example:

```ruby
Prog::Minio::MinioClusterNexus.assemble(Config.minio_service_project_id,
    "test-minio", Location::HETZNER_FSN1_ID, "minioadmin", 80, 1, 2, 1, "standard-2")
```

Results in a broken server, because it tries to use 1 drive and 2
servers. That isn't 1 pool, 2 servers per pool, and 1 drive per server
(as I assumed), it is 1 pool, 2 servers total, and 1 drive total, and
as 1/2 is 0 in Ruby, this doesn't result in a working configuration.

This adds a validation to catch this issue. I think it would better
if assemble took the per-server information, since that's what it
actually needs, but as that changes the API, it isn't included in
this commit.