Fix vm destroy workflow for VMs with load balancer
We recently updated the destroy workflow for VMs with load balancers at
fa350fa54d2d7abcd49fd2dd5093251406999211.

The workflow uses a before_run check to prevent re-running the destroy
label if it has already executed once. This check relies on the current
label. With the new changes, we need to include additional labels in
this check to ensure VMs with load balancers don't hop to destroy again.