Provision the runner again when encountering AWS insufficient capacity error
The AWS instance’s availability zone is determined by its subnet. We
could recreate the subnet and retry launching the instance.

For now, to keep things simple, I just reprovision the runner and
destroy the existing one. If we start hitting capacity issues
frequently, we can refine the logic to reprovision only the subnet
instead of the entire runner instance.