Close SSH session on event loop failure
If the event loop for the pulse check fails, close the SSH session
immediately. Otherwise, we run into an issue with Postgres server
pulse checks. These pulse checks rely on forwarding a local socket
through the SSH connection to establish a database connection.
If the SSH connection breaks in certain ways, such as the peer
resets the SSH connection, the pulse check that relies on the database
connection will hang indefinitely. The critical step is calling
`close` on the SSH object. `shutdown!` alone is not sufficient.