Fix nondeterministic failure in monitor smoke test
Instead of a single pipe, use a separate file per process.
This can fix write interleaving and partial write causes
of nondetermistic failures.

If there is still a JSON parsing problem, print the entire output
easier debugging. If there are any errors, do not delete the files,
so they can be used for analysis.