Verify VM host resources are reclaimed correctly
Recently, we had a production bug where the destroy label for a VM ran
twice, causing the VM host resources to be reclaimed twice. This
resulted in negative resource values and failed constraints. Adding a
check in the E2E tests will help verify this case. While it may not
catch every edge case, it will at least cover the basic scenario.