Allow rhizome to handle VM init scripts in setup-vm
This recognizes two parameters, init_script and init_script_args
The arguments allow for easy per-VM customization, while the
general logic is shared among multiple VMs using the init script.

For simplicity, if the parameters are not present, this treats it
the same as an empty init script with no parameters.