Add migration for vm_init_script table
This adds a foreign key in the vm table for the init script, and
a string column for the args to the init script. This allows easy
separation of per-VM data from multiple VM data

Unlike SSH public keys, We will not track per-VM init scripts, all
VM init scripts will need to be registered with the project to be
used (we can support auto registration on VM creation if we want).