Fix used_hugepages_1g calculation after VM host reboot
We verify and recalculate hugepages after rebooting the VM host by
summing SPDK hugepages and VM hugepages.

When a burstable VM is allocated on the host,
`vm_host.used_hugepages_1g` is incremented by the hugepage usage of the
slice (8g), not the burstable VM itself (4g). Therefore, if slicing is
enabled on the host, we should sum `vm_host_slice.total_memory_gib`
instead of VM memory.