Simplify and fix github object access checks
For the installation and runner, this is just simplification.

For the cache entry, the access check now requires the cache
entry being deleted is related to specific installation, as
opposed to any installation for the account. This isn't a
security issue, as there are no separate permissions per
installation, but even still, I think we should require the
cache entry be for the related installation.