Wait some time before spilling over to alien runners
According to last week’s data, waiting 10s before spilling over to alien
runners decreases the spill rate by 2x, and waiting 30s decreases it by
4x.

For now, we’ve started with a 30s wait. We plan to experiment with 10s
in the future. This approach allows us to start with lower load and
increase it later if needed.