Add spilled over vCPU capacity limit
Currently, we spill over runners to AWS EC2 instances, which naturally
have a limit on available vCPUs.

We shouldn't spill over more runners than the available vCPU capacity.

This check isn't fully concurrency-safe, but it's sufficient for now
since we don't expect too many runners to spill over simultaneously.

If we hit the limit frequently, we can consider increasing the vCPU
capacity on the AWS side.