Add new boot image for GPU VMs
Introduce a new boot image option, `gpu-ubuntu-noble`, for GPU-based VMs.
In the web interface, this image appears as `Ubuntu 24.04 for GPU VMs`.

The image is based on Ubuntu 24.04 LTS (Noble) and includes the necessary
NVIDIA drivers and tools, allowing users to leverage the GPU immediately
without additional setup.

On the VM creation page, the new boot image is only shown if
`@show_gpu != false`.