Use association instead of association dataset
Using the association dataset would have made sense if this did
the calculation in the database. However, since it didn't, it's better
to use the association, as the association can be eagerly loaded.

Without an already loaded association, it can be more efficient to
run this query in the database:

  vm.vm_storage_volumes_dataset.exclude(:boot).sum(:size_gib)

Similar approaches are possible if you are dealing with multiple
objects instead of a single object, but they increase complexity
further. We can consider that in the future if necessary.