Fix busy wait for parent strand when child strand naps in respirate
The parent would nap 0 in this case. The nap 0 makes sense for exit
or hop, since that is a state change, so immediately retrying makes
sense. However, it does not make sense for nap, since the state did
not change.

For a single child, have the parent nap using slightly less time than
the child nap, so the parent runs the child.

For multiple children, nap until slightly before the earliest
child without an active lease is scheduled.

In all cases where the child naps, do not nap less than 0.1 seconds
or more than 120 seconds.