Handle hardware_reset/reboot semaphores in more host nexus prog labels
If the host is unavailable, previously setting a reboot or hardware
reset semaphore did nothing. Do the same hardware_reset/reboot
checks in unavailable that we do in wait.

If the reboot semaphore was set, and we are stuck in reboot waiting
for the sshable to become available, allow the hardware_reset
semaphore to work.