Prepare for customer postgres subnets when destroying postgres resources
This is in preparation for making postgres private subnets be customer
subnets, and should have no current production effect. However, in the future,
customers will have the ability to add their own firewalls to the subnet,
or associate the firewall with other private subnets, and therefore should
not necessarily be deleted when deleting the postgres resource.

When deleting the customer firewall, only delete it if it has not been
renamed and it is not associated with other private subnets.

When deleting the customer private subnet, only delete it if it all of the
following are true:

* It has not been renamed

* It doesn't have any firewalls associated with it (after the firewall
  mentioned above was destroyed)

* It does not have any vms attached other than those associated to the
  postgres resource.

Additionally, dissociate the vm-specific firewalls from the postgres server
vms and destroy them.

Implement this using a new PrivateSubnet#incr_destroy_if_only_used_internally
method.