Add ips-v4 for broadcasting IP ranges we use
This is a static file in public folder for broadcasting IP ranges that
we use. This is especially useful for GHA users to allowlist their endpoints
for requests from Ubicloud runner VMs.

The format is the same with CloudFlare's: https://www.cloudflare.com/ips-v4/

We should remember to update the file as we create new hosts:

```ruby
Address
.where { (family(cidr) =~ 4) & (masklen(cidr) !~ 32) }
.map { NetAddr::IPv4Net.new(_1.cidr.network, NetAddr::Mask32.new(16)).to_s }
.uniq
.sort
.each { puts _1 };
```