Add ips-v4 for broadcasting IP ranges we use
This change adds a new endpoint /ips-v4 that reports the IP ranges we can use.
This is especially useful for GHA users to allowlist their endpoints
for requests from Ubicloud runner VMs. The results are calculated initially
once and reused on subsequent calls.

The format is the same with Cloudflare's: https://www.cloudflare.com/ips-v4/