Fix passed frame for load balancer buds
In our codebase, we try to keep our JSON keys as string, since when they
fetched from DB, they are always string. When we pass it as a symbol, it
causes failure especially in equality check.

I found a place where we are passing symbol instead of string thanks to
new version of json package. It prints following warning:

    This will raise an error in json 3.0 unless enabled via `allow_duplicate_key: true`
    .../mise/installs/ruby/3.4.7/lib/ruby/3.4.0/delegate.rb:349: warning: detected duplicate key "subject_id" in {"subject_id" => "9749a267-586c-842b-962c-666f22da28fa", subject_id: "adede52c-c92e-8774-9b28-e33ff0c42d62", load_balancer_id: "9749a267-586c-842b-962c-666f22da28fa"}.

As you see `"subject_id"` and `:subject_id` are different keys. They
can't merged in `hash.merge` method.