Configure memory_overcommit
When OOM killer kills one of the PostgreSQL processes, there is risk of process
leaving the shared memory in an inconsistent state. To prevent this, when one
of its processes is killed, postmaster intentionally crashes itself and then
performs crash recovery on the next startup. This is required to ensure data
consistency.

Still, it is not a good experience to have PostgreSQL crash due to OOM killer,
so we configure memory overcommit to be strict.