Do not show VMs in different project when viewing private subnets in web UI
Only show nic related information if the nic doesn't have an associated VM,
or the VM is in the same project as the private subnet. This will make it so
that Ubicloud-managed VMs located in customer subnets do not show up as
attached VMs on the private subnet's "Virtual Machines" page.