Add extra fields about postgres connection info
We normally only return the connection string, but then clients might need to
parse it if they need individual fields, which is not very straightforward as
each field might contain special characters requiring extra attention during
parsing. So we are adding username, password and hostname fields to make it
easier for clients to consume this information.