Change networking for new postgres resources
Before, postgres resources were placed in a private subnet in the
postgres project, with a firewall in the postgres project, and
extra firewall rules could be added using postgres-specific
PostgresFirewallRules.

With this approach, postgres resources are created in a private
subnet in the customer's project, with a firewall in the customer's
project controlling the customer's desired access.  To ensure the
managed service still works regardless of customer firewall changes,
an internal firewall in the postgres project is added to each
separate postgres server VM.

This results in a temporary situation where some postgres resources
have an internal firewall (new approach), and some do not (old
approach). Future changes will convert the old approach to the new
approach by adding an internal firewall and updating existing
postgres servers to use it. Once that change is made, some of the
new branches in this commit will be removed.