Log spot instance interruptions for alien runners
Spot instances are cost-effective but can be interrupted by AWS when
capacity is reclaimed.

When this happens, SSH commands begin to fail. In such cases, we check
the termination status via the AWS API.

For now, we simply log these interruptions. In the future, we may stop
spilling over to alien runners when the number of interruptions becomes
high.