Fix Prog::Kubernetes::KubernetesClusterNexus#destroy to handle worker nodes when checking the subnet
To test this, fix the Prog::Kubernetes::KubernetesClusterNexus specs to test
more realistic cases, using Prog::Kubernetes::KubernetesClusterNexus.assemble.
This ensures the private subnet and firewalls are setup as expected, and that
objects have strands so you can test for semaphores.

Test that the private subnet is deleted when expected, including when there
are worker nodes and not just cp nodes.

Test that private subnet is not deleted if there are other resources in the
private subnet.