Avoid unnecessary queries for when there is a lack of capacity
Move utilization inside the block where it is used.

For queued_vms, this makes the non-github-runners case faster by
doing a count in the database instead of retrieving all records
in Ruby.

For the github-runners case, this potentially does two queries
instead of one, but the queries are much simpler. Instead of
getting all rows, there is one query for the count, and a
separate query that just returns the ids.