Recursively reap children in Prog::Base#reap
If you have a child strand that exited, and that strand has an
unreaped child strand that exited, then before this change, reap
will fail with a foreign key violation, because it tries to destroy
the child strand while the grandchild strand still exists.

While this case is unlikely, it is possible to hit for at least
kubernetes:

* KubernetesNodepoolNexus#destroy reaps
  * Has ProvisionKubernetesNode child with exitval
    * Which has BootstrapRhizome child with exitval

ProvisionKubernetesNode in this case didn't reap the
BootstrapRhizome child because before the reaping, the related
cluster was destroyed.