Handle customer destroying/detaching the postgres resource specific firewall
In the web UI networking page, include a link to the private
subnet networking page.

If there are attempts to access/modify firewall rules, return an
error.

This does not change the API response for getting the postgres
resource. That will still show the PostgresFirewallRule entries
even if they are no longer used.