Add Prog for migrating SPDK disks to Ubiblk
This prog first attempts to stop the VM and its spdk controller,
creating the required configs for runnning the migration.

The migration happens using a prebuilt binary from the ubiblk-migrate
project. It utilizes damonizer2 to run the migration since it takes a
long time to convert the disk.

After finishing up the migration, it starts the VM again and sets
the Vms strand back to the wait label.