Make postgres GET firewall-rule route use customer firewall rules
Previously, this used the PostgresFirewallRule related to the
PostgresResource. With the switch to using customer firewalls
and customer subnets for PostgresResources, PostgresFirewallRule
is an unnecessary abstraction.

Add port to API responses for postgres firewall rules. This previously
wasn't needed because each postgres firewall rule was for 2 ports, but
now that each port is handled by a separate rule, the port is needed to
disambiguate.  Temporarily, use 5432 if it is a PostgresFirewallRule,
but that will be going away shortly.

If the customer firewall allows other ports, they will not be
shown in this view. This is designed for keeping as much
backwards compatibility as possible while removing
PostgresFirewallRule. For full information on firewall rules,
the private subnet for the Postgres database should be used.
Add PostgresResource#pg_firewall_rules to handle this filtering.

To avoid duplicate queries, have postgres_require_customer_firewall!
helper return the firewall if it is found.

Temporarily, have the openapi.yml configuration allow either
pf or fr ubids for PostgresFirewallRule. This will eventually be
switching to all fr when the conversion is complete.