Make postgres PATCH/DELETE firewall-rule route work with FirewallRule ubids
Add PostgresResource#pg_firewall_rule{,s_dataset} to make it easier
to retrieve a specific rule.

Make PrivateSubnet#update_private_subnet_firewall_rules public,
so it can be called in the routes.

Update openapi.yml to remove use of pf ubids. The
postgres_firewall_rule_id component can be switched to firewall_rule_id.

This removes the last usage of PostgresFirewallRule in the routes.
The serializer still refers to PostgresFirewallRule, but it operates
on FirewallRule instances only now.