Allow text component to take id and handle_params_and_errors keyword arguments
Previously, the text component was not able to correctly handle multiple inputs
with the same name on the same page, because it would use the same id for each,
and that is not allowed. Support multiple inputs with same name by having an
id keyword argument that defaults to name.

Error handling when using multiple inputs with the same name on the same page
was similarly broken. Accept a handle_params_and_errors keyword argument (true
by default) so that we can suppress overriding the value and setting an error.
Expected usage when using the multiple inputs with the same name on the same
page is that is only the inputs affected by a form submission would have
handle_params_and_errors set.