Support modifying firewall rules in web UI and API/SDK/CLI
This is loosely based on the postgres firewall rule modification
support.

This introduces a new pattern, where web paths do not exactly match
api paths. The API uses PATCH (as postgres firewall rules did), but
the web uses POST. This allows the web portion to work without
javascript, so normal testing with capybara works. I think it would
be good to expand the usage of this pattern to the rest of the web
routes and views, but that's a future change.

Inside of an inline edit component (which requires javascript), just
use text boxes for each input and normal form submission buttons.