Support adding a firewall rule using private subnet name or id instead of cidr in the API/SDK/CLI
When adding using private subnet name or id, this creates 2 rules, one
for IPv4 and one for IPv6. This changes the response to include multiple
rules if a private subnet name or id is given. Since it wasn't possible
to use a private subnet name or id before, this is in general backwards
compatible. However, code that passed values through without checking
them (such as the CLI) needs adjustment. I think the amount of backwards
compatibility breakage here is acceptable.