Add BYOC (AWS) dummy location selector
Since we have recently introduced Ubi PG on AWS and we are capable of
running inside the customer account, we are adding a dummy location
selector to the console UI. This way, we will point the interested
parties and allow list manually. It will require AMI sharing and
location creation using PrivateLocation.