Avoid Prog::Test2 superclass mismatch error when reloading in development environment
Nest Test2 under Prog::Test so that when Prog::Test is reloaded,
there is no superclass mismatch. This is only used in one test.
I'm not sure why the spec uses a subclass, since it appears to be
testing stacks and not subclass-specific behavior, but this seems
to be a simple way to fix the issue.