Use the stack information and fix the loadbalancing by unique port
We start using the stack information while configuring load balancing.
This way, if the ipv4 stack is up on a node while the ipv6 is down, it
keeps getting the traffic for ipv4 but not for ipv6.

More importantly, we are fixing a sneaky bug here. Previously, using
uniq on load_balancer.active_vm_ports collapsed multiple
load_balancer_vm_ports that shared the same load_balancer_port. This
caused traffic to be routed to only one node instead of multiple.
In some cases, it could even select the active VM’s own port, resulting
in an empty neighbor set and a total routing failure. The uniq property
was added to avoid an undeterministic bug which we solve here by sorting
on the stack column.