Store bank transfer info if no payment method
If a project has billing information but no payment method, it means we
collect payments via bank transfer. In this case, we need to store the
bank transfer details since we use different accounts for EU and non-EU
customers, and they might change over time. This commit stores the
information in the invoice; the next commit will display it in the
invoice PDF.