Change ubid type of BillingInfo from `et` to `b1`
While working on something else, I realized that the ubid type
of BillingInfo was set to ETC_TYPE in #3360. So all BillingInfo's
after the change have their ubids start with `et` not `b1` and
causing decoding to break.

This change also adds a test that check all types (with exceptions)
have the correct ubid_type defined.