Fix Ubi PG on AWS VPC creation with tcp 22 allow rule
With a recent change, we moved the tcp 22 rule into an internal firewall
which are attached to VMs once they are up and running. However, AWS
security groups are created through VPCs which are part of
private_subnets. Therefore, not having the tcp 22 rule on the
private_subnet firewall breaks the initial VM provisioning in AWS.
Because, when we reach wait_sshable, the security group lacks the tcp 22
allow rule since it is part of the vm specific rules.  Those are applied
only after we switch to wait state.  This is also necessary because a
customer may want to have subnets that are completely locked down and
have no allow rules in their firewall. In that case, we cannot verify
sshability at the provisioning time if we process the firewalls before.