Add one load_balancer_vm_port for each stack option
Since the load balancer health checks are performed through
load_balancer_vm_port entity, and I want to separate the health state of
each stack for the same VM, I am adding a new row per stack.
This way, if the ipv4 or ipv6 (single) endpoint is unhealthy for a dual
stack load balancer, we will be able to individually identify that.

This commit simply populates the rows for new instances load balancer
instances.