Fix strand test with undeterministic mock count
This test mysteriously fails on my local machine even on the main
branch, but it seems to pass on CI. It might be a macOS specific issue.

    Failures:

      1) Strand logs end of strand if it took long
        Failure/Error: expect(Time).to receive(:now).and_return(Time.now - 10, Time.now, Time.now)

          (Time (class)).now(*(any args))
              expected: 3 times with any arguments
              received: 4 times with any arguments
        # ./spec/model/strand_spec.rb:95:in 'block (2 levels) in <top (required)>'
        # ./spec/spec_helper.rb:60:in 'block (3 levels) in <top (required)>'
        # ./spec/spec_helper.rb:59:in 'block (2 levels) in <top (required)>'

    Finished in 46.92 seconds (files took 6.28 seconds to load)
    4108 examples, 1 failure

    Failed examples:

    rspec ./spec/model/strand_spec.rb:92 # Strand logs end of strand if it took long

Since we are generally loose on the mock count and don't specify an
exact number, we can loosen this test by using `.at_least(:once)` as
well.

Interestingly, we're mocking `Time.now`, but the mock returns the actual
`Time.now` again.