Spread Kubernetes Control Plane VMs Among Hosts
We want to avoid CP VMs landing on the same hosts. This change
attempts to do that by utilizing exclude_host_ids parameter of
VM assembler.

This could lead to a small problem, when replacing the nodes in an
upgrade scenario where we don't necessarily want to avoid the new
nodes to land on the same host, but that's a minor inconvenience
for now.

This is step #1 of improvements on the node distribution and later
we can implement some affinity-like levers in allocator to get
better spread with AZs or worker nodes as well.