Add support for updating firewall descriptions
Previously, we didn't offer a way to do this. So if you wanted to change
the description after creating the firewall, you were out of luck.

Mostly by Claude Code, using the following prompt:

```
Ubicloud currently supports updating the name of the firewall but
not updating the description of a firewall. Add support for updating
a description of the firewall, similar to the current support for
updating the name for a firewall. Steps:

* Add an appropriate endpoint to routes/project/location/firewall.rb to change the firewall description
  * To allow API requests to the route to work, update the OpenAPI configuration stored in openapi/openapi.yml
* Update views/networking/firewall/settings.erb template to add a form that will submit to the new endpoint.
  * Add appropriate capybara-based specs to test this change to spec/routes/web/firewall_spec.rb
* Update sdk/ruby/lib/ubicloud/model/firewall.rb to add support for calling that route to the SDK
* Add cli-commands/fw/post/update-description.rb to add support to the CLI, using the SDK support
  * Add appropriate specs to test this change to a new file named spec/routes/api/cli/fw/update-description_spec.rb, similar to the existing specs in spec/routes/api/cli/fw/rename_spec.rb
```

Issue fixed manually:

* Fixed the CLI command by including `UbiCli.` before `on`
* Change r`request.redirect` to `r.redirect` in the route
* Add test for case where description is unchanged
* Use different flahs notice when description is unchanged
* Remove unused variable assignment in spec
* Remove redundant self in SDK method.
* Change CLI command description
* Update golden files
* Use PATCH instead of POST for the API, similar to other
  objects

This uses r.api_patch_web_post so that the web UI uses POST,
allowing for normal Capybara testing. This required moving the
route above the r.is call.