Proxy ARP on all interfaces for IPv4
We used the enable proxy_arp per interface during VM provisioning.
However, that has a drawback. The proxy_arp entries do not stick when an
interface is swapped. For example, after a recent incident, our provider
had to replace the server chassis which resulted in the interface names
to be refreshed. This meant, our proxy_arp settings to be refreshed as
well. With this setting to be applied with "all", we don't need to
refresh the proxy_arp setting anymore.