Use next minute for backup target times
Our datetime picker omits seconds, so when we try to use the exact backup time,
we round it down to the earlier minute. The problem is that time is the worst
possible time to pick as restore target because it requires applying entire day
worth of WAL changes. By rounding up to the next full minute, we encourage
people to pick restore targets that are faster to complete.