Add allow_unescaped web helper method
This method just returns the argument. Update the XSS scanner to not
flag calls to allow_unescaped.

This is currently used for places where we are expecting HTML, but
don't want to add a bunch of additional scan patterns to the rake task.
It may be better to rename the variables to end in _html (as was done in
an earlier commit), but these are places where doing so may be a
substantial amount of work.