Move blob_storage finding logic to timeline model
In PostgresTimelineNexus::assemble we used to find and set the blob_storage_id
while creating the PostgresTimeline object. However there is no blob_storage_id
for AWS resources. So it doesn't make sense to have this logic so early where it
might not be even needed at all. Instead we are pushing this logic down to the
model where we actually need to access the blob storage.